Changes to Network Message Buffering & Removing Message Send Queue
Changed NetworkMessage from a class to a struct to ease the load on the garbage collector.
Removed the MessageSendQueue as it was causing memory leaks by allocating network streams and not recycling properly, I believe the issue was that I was not clearing message send queues for clients when they disconnect, non the less, the whole idea of a message send queue just seems inefficient, too much fragile code for too little gain, so now I just send every message directly without queuing it, LiteNetLib already queues packets internally so it doesn't need it, web-socket-sharp doesn't do this but it's the web after all, you shouldn't expect stellar performance out of anything web.