Adjusting Network Payload Strategy
Completely removed NetworkMessages concept, payloads are now sent directly with a type prefix to distinguish between them.
This helps levitate some of the stress on the garbage collector because now payloads are read generically on the spot where they are needed.
The message buffer system was also changed, allowing to register any type of payload and retrieving a handle when doing so.